+++ source/news.rst
@@ -1,9 +1,26 @@
 News & Updates
 ==============
 
-03 August 2026
---------------
+..  news-item:: New atom feed
+    :date: 2026-08-06
 
-`Diátaxis is now available in Polish <https://diataxis.fr/pl>`_, courtesy of the translation efforts of `Piotr Nawrot <https://piotrnawrot.com/>`_. Other transations are in progress.
+    Added an atom feed, https://diataxis.fr/atom.xml.
 
-:doc:`More information about translating Diátaxis <translation>`
+..  news-item:: Polish translation
+    :date: 2026-08-04
+
+    `Diátaxis is now available in Polish <https://diataxis.fr/pl>`_, courtesy of the translation efforts of `Piotr Nawrot <https://piotrnawrot.com/>`_.
+
+    Other translations are in progress.
+
+    :doc:`More information about translating Diátaxis <translation>`
+
+Heading
+-------
+
+Date
+
+Entry
+marker 1785963858
+
+.. comment 1785963887.4449122